separate HighwaySystem.h
