+++ siteupdate/cplusplus/siteupdate.cpp
@@ -51,6 +51,7 @@ class HGEdge;
 #include "classes/ConnectedRoute/ConnectedRoute.h"
 #include "classes/Waypoint/Waypoint.h"
 #include "classes/HighwaySegment/HighwaySegment.h"
+#include "classes/HighwaySystem/HighwaySystem.h"
 #include "classes/TravelerList/TravelerList.h"
 #include "classes/GraphGeneration/HGEdge.h"
 #include "classes/GraphGeneration/PlaceRadius.h"
@@ -68,7 +69,7 @@ class HGEdge;
 #include "classes/ElapsedTime.cpp"
 #include "classes/ErrorList.cpp"
 #include "classes/Region.cpp"
-#include "classes/HighwaySystem.cpp"
+#include "classes/HighwaySystem/HighwaySystem.cpp"
 #include "classes/DatacheckEntry.cpp"
 #include "classes/DatacheckEntryList.cpp"
 #include "classes/Waypoint/Waypoint.cpp"
